requirements.txt
# Web UI
streamlit>=1.30.0

# Resume parsing
pdfplumber>=0.10.0
python-docx>=1.0.0

# NLP
spacy>=3.7.0

# Machine learning / matching
scikit-learn>=1.3.0

# Job scraping
python-jobspy>=1.1.0

# Environment management
python-dotenv>=1.0.0

# Scheduling
APScheduler>=3.10.0

# AI cover letter (optional)
openai>=1.0.0

# PDF resume generation
reportlab>=4.0.0

# Browser automation (future use)
playwright>=1.40.0

# Testing
pytest>=7.4.0

# Authentication
bcrypt>=4.0.0
